Refatorando código do backend
